fixed a test :)
